Add files via upload
This commit adds the files for the SIH project developed under the Gokul Mission. The project focuses on image-based animal type classification for cattle and buffaloes. It uses machine learning and web technologies to identify animal features, extract body parameters, and generate reliable classification results. The system aims to reduce human error, support farmers, and integrate with the Bharat Pashudhan App to improve accuracy and efficiency in livestock management.